requirements.txt
streamlit>=1.32.0
langchain-community>=0.0.21
langchain-core>=0.1.27
langchain-groq>=0.1.0
crewai>=0.16.2
youtube-dl>=2021.12.17
pytube>=15.0.0
pysqlite3-binary>=0.5.2
groq>=0.4.0
youtube-transcript-api>=0.4.0
transformers
langchain
langchain_deepseek